+++ CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## UNRELEASED
 ### Changed
 - renamed LanguagesCollection and SnippetsCollection
+- rewritten synchronization snippets logic
 ### Added
 - LanguagesCollection can be searched by code
 - LanguagesCollection can be searched by file name extension
 - Managed schema version with db drop and create
+- Added "dirty" and "deleted" fields on snippet model
 
 ## [1.1.1] - 2017-04-22
 ### Changed